NIAD-509 Sandbox refactoring
