+++ docker/service/src/integration-test/java/uk/nhs/adaptors/scr/component/ScrControllerTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.nhs.adaptors.scr.clients.SpineClient;
+import uk.nhs.adaptors.scr.clients.SpineClientContract;
 import uk.nhs.adaptors.scr.components.FhirParser;
 import uk.nhs.adaptors.scr.config.ScrConfiguration;
 import uk.nhs.adaptors.scr.config.SpineConfiguration;
@@ -48,7 +48,7 @@ public class ScrControllerTest {
     private ScrConfiguration scrConfiguration;
 
     @MockBean
-    private SpineClient spineClient;
+    private SpineClientContract spineClient;
 
     @Test
     public void whenRequestProcessingTakesTooMuchTimeExpect504() {
