+++ docker/service/src/main/java/uk/nhs/adaptors/scr/config/ScrSpringConfiguration.java
@@ -3,9 +3,9 @@ package uk.nhs.adaptors.scr.config;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import uk.nhs.adaptors.scr.clients.MockSpineClient;
-import uk.nhs.adaptors.scr.clients.ProdSpineClient;
+import uk.nhs.adaptors.scr.clients.SandboxSpineClient;
 import uk.nhs.adaptors.scr.clients.SpineClient;
+import uk.nhs.adaptors.scr.clients.SpineClientContract;
 import uk.nhs.adaptors.scr.clients.SpineHttpClient;
 
 @Configuration
@@ -16,14 +16,11 @@ public class ScrSpringConfiguration {
 
     @Bean
     @Autowired
-    public SpineClient spineClient(SpineConfiguration spineConfiguration, SpineHttpClient spineHttpClient) {
-        switch (scrConfiguration.getSpineMode()) {
-            case REAL:
-                return new ProdSpineClient(spineConfiguration, spineHttpClient);
-            case MOCK:
-                return new MockSpineClient();
-            default:
-                throw new IllegalStateException("Invalid scr.spineMode variable value: " + scrConfiguration.getSpineMode());
+    public SpineClientContract spineClient(SpineConfiguration spineConfiguration, SpineHttpClient spineHttpClient) {
+        if (scrConfiguration.getSandboxMode()) {
+            return new SandboxSpineClient();
+        } else {
+            return new SpineClient(spineConfiguration, spineHttpClient);
         }
 
     }
