+++ docker/service/src/main/java/uk/nhs/adaptors/scr/services/AcsService.java
@@ -3,7 +3,7 @@ package uk.nhs.adaptors.scr.services;
 import com.github.mustachejava.Mustache;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import uk.nhs.adaptors.scr.clients.SpineClient;
+import uk.nhs.adaptors.scr.clients.SpineClientContract;
 import uk.nhs.adaptors.scr.clients.SpineHttpClient;
 import uk.nhs.adaptors.scr.models.ACSPayload;
 import uk.nhs.adaptors.scr.utils.TemplateUtils;
@@ -14,7 +14,7 @@ import java.util.Map;
 @Component
 public class AcsService {
     @Autowired
-    private SpineClient spineClient;
+    private SpineClientContract spineClient;
 
     private static final Mustache SET_RESOURCE_PERMISSIONS_TEMPLATE = TemplateUtils.loadTemplate("set_resource_permissions.mustache");
